Keep a logger file in sparseml.pytorch.utils
