+++ src/sparseml/core/__init__.py
@@ -20,6 +20,7 @@ from .factory import *
 from .framework import *
 from .framework_object import *
 from .lifecycle import *
+from .logger import *
 from .model import *
 from .modifier import *
 from .optimizer import *
